# [perf] Use color triplets