--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,3 +6,10 @@ authors = ["Tony Crisci <tony@dubstepdish.com>"]
 [dependencies]
 vpsearch = "1.3.4"
 lazy_static = "1.0"
+
+[dev-dependencies]
+criterion = "0.1.2"
+
+[[bench]]
+name = "basic"
+harness = false