--- a/src/colors/roygbiv.rs
+++ b/src/colors/roygbiv.rs
@@ -1,5 +1,5 @@
 use color_names::ColorNames;
-use color::{Color, color_from_hex};
+use color::{Color, color_from_triplet};
 
 pub struct RoygbivColors {
 }
@@ -7,13 +7,13 @@ pub struct RoygbivColors {
 impl ColorNames for RoygbivColors {
     fn get_colors(&self) -> Vec<Color> {
         return vec![
-            color_from_hex("red", "#FF0000"),
-            color_from_hex("orange", "#FFA500"),
-            color_from_hex("yellow", "#FFFF00"),
-            color_from_hex("green", "#008000"),
-            color_from_hex("blue", "#0000FF"),
-            color_from_hex("indigo", "#4B0082"),
-            color_from_hex("violet", "#EE82EE")
+            color_from_triplet("red", (255, 0, 0)),
+            color_from_triplet("orange", (255, 165, 0)),
+            color_from_triplet("yellow", (255, 255, 0)),
+            color_from_triplet("green", (0, 128, 0)),
+            color_from_triplet("blue", (0, 0, 255)),
+            color_from_triplet("indigo", (75, 0, 130)),
+            color_from_triplet("violet", (238, 130, 238))
         ]
     }
 }