… src/colors/html.rs
… src/colors/ntc.rs
… src/colors/pantone.rs
… src/colors/x11.rs