Ran a benchmark before making changes. It's just testing the example from the readme.
```
basic                   time:   [99.085 us 99.918 us 100.82 us]
```

Then replaced the color parsing with hard coded integers, and ran the benchmarks again.

```
basic                   time:   [13.337 us 13.388 us 13.443 us]
                        change: [-86.708% -86.590% -86.469%] (p = 0.00 < 0.05)
                        Performance has improved.
```